+++ package.json
@@ -9,7 +9,7 @@
     "react-dom": "^16.12.0",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
+    "react-scripts": "3.4.1",
     "redux": "^4.0.5"
   },
   "scripts": {
@@ -41,6 +41,11 @@
     "@testing-library/react": "^9.5.0",
     "cypress": "^3.7.0",
     "cypress-multi-reporters": "^1.4.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-react": "^7.20.6",
+    "eslint-plugin-react-hooks": "^4.0.0",
     "mocha": "^7.2.0",
     "mochawesome": "^6.1.1",
     "mochawesome-merge": "^4.0.3",
