+++ src/mock_data/categories.js
