--- src/pages/GamePage.jsx
@@ -1,17 +1,0 @@
-import React, { Component } from 'react';
-// import Header from '../components/Header';
-// import Quizz from '../components/Quizz';
-
-class GamePage extends Component {
-    render() {
-        return (
-            <div>
-                {/* <Header />
-                <Quizz />                 */}
-                «Perguntas»
-            </div>
-        )
-    }
-}
-
-export default GamePage
