--- src/pages/SettingPage.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-class SettingPage extends Component {
-    render() {
-        return (
-            <div>
-                «Configurações»                
-            </div>
-        )
-    }
-}
-
-export default SettingPage;
