+++ src/pages/SettingsPage.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class SettingsPage extends Component {
+  render() {
+    return (
+      <div>
+        «Configurações»
+      </div>
+    );
+  }
+}
+
+export default SettingsPage;
