+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import loginReducer from './login';
+
+const rootReducer = combineReducers({
+  loginReducer,
+});
+
+export default rootReducer;
