--- src/reducers/index.jsx
@@ -1,8 +1,0 @@
-import {combineReducers} from 'redux';
-import loginReducer from './login';
-
-const rootReducer = combineReducers({
-    loginReducer
-})
-
-export default rootReducer;
